fixed some membership api
git-svn-id: http://svn.civicrm.org/civicrm/trunk@7847 cff41f8d-9fe8-0310-9d4e-fb156a3ffe5e

+++ CRM/Member/BAO/MembershipType.php
@@ -155,6 +155,10 @@ class CRM_Member_BAO_MembershipType extends CRM_Member_DAO_MembershipType
         require_once 'CRM/Member/DAO/MembershipType.php';
         $membershipType =& new CRM_Member_DAO_MembershipType( );
         $membershipType->id = $membershipTypeId;
+        //fix for membership type delete api
+        if (! $membershipType->find(true )) {
+            return new CRM_Core_Error("Unable to find the membership type");
+        }
         $membershipType->delete();
     }
